age_estimation/requirements.txt
numpy==1.16.0
pytorch-ignite==0.1.2
torch==1.0.0
torchvision==0.2.1
pillow==5.3.0